switch what TIterable means, define PIterable, have PContainer/PStream inherit from that
