+++ hail/src/main/scala/is/hail/expr/types/physical/PIterable.scala
@@ -1,0 +1,5 @@
+package is.hail.expr.types.physical
+
+abstract class PIterable extends PType {
+  def elementType: PType
+}
