+++ hail/src/main/scala/is/hail/expr/types/virtual/TIterable.scala
@@ -1,11 +1,12 @@
 package is.hail.expr.types.virtual
 
-import is.hail.annotations._
+import is.hail.expr.types.physical.PIterable
+import is.hail.utils.FastSeq
 
-abstract class TIterable extends TContainer {
-  override def valuesSimilar(a1: Annotation, a2: Annotation, tolerance: Double, absolute: Boolean): Boolean =
-    a1 == a2 || (a1 != null && a2 != null
-      && (a1.asInstanceOf[Iterable[_]].size == a2.asInstanceOf[Iterable[_]].size)
-      && a1.asInstanceOf[Iterable[_]].zip(a2.asInstanceOf[Iterable[_]])
-      .forall { case (e1, e2) => elementType.valuesSimilar(e1, e2, tolerance, absolute) })
+abstract class TIterable extends Type {
+  def physicalType: PIterable
+
+  def elementType: Type
+
+  override def children = FastSeq(elementType)
 }
