+++ hail/src/main/scala/is/hail/expr/types/virtual/TSet.scala
@@ -8,7 +8,7 @@ import org.json4s.jackson.JsonMethods
 
 import scala.reflect.{ClassTag, classTag}
 
-final case class TSet(elementType: Type, override val required: Boolean = false) extends TIterable {
+final case class TSet(elementType: Type, override val required: Boolean = false) extends TContainer {
   lazy val physicalType: PSet = PSet(elementType.physicalType, required)
 
   override val fundamentalType: TArray = TArray(elementType.fundamentalType, required)
